Detail the expected format of async env file
The docs were a bit unclear about what the format of an asynchronous env file should be. Looking into the code I see that the file should export (rather than return) either an object or a Promise resolving to an object. So I updated the readme to be clearer about this.